requirements.txt
torch
openai
pydantic
whisperx
pysrt
pydub
pycountry
pyaudio
speechmatics-rt